split logic out
